requirements.txt
asgiref==3.5.0
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
black==22.1.0
click==8.0.3
decorator==5.1.1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
executing==0.8.2
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nltk==3.7
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.7
pytesseract==0.3.9
pytz==2021.3
regex==2022.3.15
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
toml==0.10.2
tomli==2.0.0
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
